+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/entities/BeeInteractivity.java
@@ -214,7 +214,7 @@ public class BeeInteractivity {
             playerEntity.addEffect(new MobEffectInstance(
                     BzEffects.PROTECTION_OF_THE_HIVE.get(),
                     BzBeeAggressionConfigs.howLongProtectionOfTheHiveLasts,
-                    2,
+                    0,
                     false,
                     false,
                     true));